pyproject.toml
[project]
name = "dating-profile-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.42.0",
    "python-dotenv>=1.0.1",
]
